[PHASE-G] fix: 設定選單捲動視窗接上真實 offset，修正游標移出可見範圍不顯示
2026-09-02 對 Task 2 最終 commit（c980927）補跑 codex Tier 3 review 抓到
3 個 CRITICAL（同一根因）：wrapSettingsCursor() 用 SETTINGS_MENU_ITEM_COUNT
（該次從 5 改成 8）做 wrap 邊界，是 inline 定義在 header 的純函式，
input_handler.cpp 完全沒改一行就自動繼承新邊界——UP/DOWN 早就能把
settingsCursor 移到 5、6、7（App連線設定／Type-C連線／裝置資訊），但
main.cpp:1177 的正式渲染呼叫點寫死 scroll_offset=0。實機上使用者按 DOWN
超過第 4 項，游標其實有移動，但畫面高亮直接消失，要按到 wrap 回 0 才恢復。

改動（僅修這個 CRITICAL，不擴大到 Task 3 完整範圍——cursor 5~7 的
BTN_PRIMARY 分派〔進入 App連線/Type-C連線/裝置資訊子畫面〕仍留給 Task 3）：
- 新增 settingsScrollOffset 全域（main.cpp／app_globals.h／input_handler.cpp
  extern），跟 settingsCursor 一樣 sticky、不在進入選單時重設——兩者永遠
  成對更新，拆開重設會讓其中一個落單，重新製造同一個 bug
- 新增 advanceSettingsCursorAndScroll()（ui_settings.h）：把 wrapSettingsCursor()
  + clampScrollOffset() 的串接合併成一次呼叫回傳新的一對值，input_handler.cpp
  的 BTN_UP/BTN_DOWN 分支改用這個函式，讓呼叫端拿到的是已配對好的結果，沒有
  機會只更新其中一個——這是本輪 review 的 tests 面向抓到的 IMPORTANT（這段
  成對更新邏輯先前完全沒有測試涵蓋，因為 input_handler.cpp 屬 src/，native
  build 排除），抽成純函式後在 test_settings_ui 補了 5 個涵蓋視窗內移動／下捲
  觸發／連續下捲到底／wrap 回頂端／wrap 到底端的測試
- main.cpp:1177 渲染呼叫改傳 settingsScrollOffset，不再寫死字面值 0
- drawSettingsMenu() 的 scroll_offset 參數型別從 uint8_t 改成 uint16_t，
  對齊 settingsScrollOffset 實際型別，消除隱性窄化轉型（codex types 面向、
  agent code-review 與 rules 三個獨立審查者都各自抓到同一個窄化）
- DisplaySnapshot／DisplaySnapshotInputs／captureSnapshot 補上
  settingsScrollOffset 欄位（同類回歸的又一次，見 historyScrollOffset 的
  既有 pattern）；test_display_snapshot.cpp 補對應 regression case
- 附帶修正 ui_settings.h 的 scroll_offset JSDoc（原本宣稱「main.cpp 呼叫點
  一律傳字面值 0，接線在 Task 3」，現已不實）；修正 input_handler.cpp 內
  誤導性的 commit 歸因註解（原寫「c980927 CRITICAL」，暗示該次 review 就抓到
  這個問題，實際上是事後補跑才發現，comments 面向抓到）

未修：codex/agent 同批抓到的其餘 IMPORTANT/MINOR——settingsCursor/
settingsScrollOffset 這對不變量目前只靠慣例維持、建議 Task 3 抽成共用
ScrollCursorState 型別（跟 historyCursor/historyScrollOffset 是同一種
待統一模式）；恢復預設對話框跟捲動後的游標高亮格重疊（c980927 既有小 bug，
本次讓觸發面從 1/8 擴大到 4/8，純視覺，不影響功能）——維持既有殘餘風險記錄，
不在本次最小修復範圍內處理。

驗證：native test 643 cases 全綠（含新增 6 個 regression/pairing test），
唯一既有 ERRORED 為 test_storage_hw（與本次無關）；ESP32 build SUCCESS
Flash 72.8%

Review 方法說明：codex CLI 服務端在本輪同樣出現持續性 capacity 錯誤，經
使用者同意全部 6 個面向改派 Claude subagent 執行（pr-reviewer lite +
pr-review-toolkit 五個 agent），六個面向皆已收齊、無降級。

+++ firmware/lib/ui_settings/ui_settings.cpp
@@ -143,7 +143,7 @@ static_assert(SETTINGS_MENU_ITEM_COUNT == SETTINGS_MENU_COUNT,
   * @param device_name_locked  裝置名稱是否鎖定（由呼叫端算好）
   * @param restore_confirm     恢復預設確認對話框是否顯示中
   */
- void drawSettingsMenu(Display& disp, uint8_t cursor, uint8_t scroll_offset,
+ void drawSettingsMenu(Display& disp, uint8_t cursor, uint16_t scroll_offset,
                        bool device_name_locked, bool restore_confirm) {
      // STEP 01: 讀取裝置名稱（ESP32: LittleFS / native: mock FS）
      //   注意：此處不得重設 s_brightness——開機時 main.cpp setup() 已用 NVS 值灌入，
